+++ misc/systemd_explorer/systemd_install.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+install ./*.service /usr/lib/systemd/system
+systemctl reload-daemon
+systemctl enable pyborg_*
+journalctl