+++ pyborg/pyborg/commands.py
@@ -2,33 +2,33 @@ import logging
 
 import requests
 
-from .util.irc import command
-
 import pyborg
+
+from .util.commands import command
 
 logger = logging.getLogger(__name__)
 
 
 @command()
-def info():
+def info() -> str:
     """Returns version number and source code link"""
     return "I am a version {} Pyborg. My source can be found at https://github.com/jrabbit/pyborg-1up".format(pyborg.__version__)
 
 
 @command(internals=True)
-def words(multiplex, multi_server):
+def words(multiplex: bool, multi_server: str) -> str:
     """Returns the number of words known and contexts per word"""
     if multiplex:
         ret = requests.get(multi_server + "words.json")
         ret.raise_for_status()
-        words = ret.json()
+        payload = ret.json()
         try:
-            contexts_per_word = float(words["contexts"]) / float(words["words"])
+            contexts_per_word = float(payload["contexts"]) / float(payload["words"])
 
         except ZeroDivisionError:
             contexts_per_word = 0
 
-        msg = "I know %d words (%d contexts, %.2f per word), %d lines." % (words["words"], words["contexts"], contexts_per_word, words["lines"])
+        msg = "I know %d words (%d contexts, %.2f per word), %d lines." % (payload["words"], payload["contexts"], contexts_per_word, payload["lines"])
         return msg
 
     else:
@@ -36,7 +36,7 @@ def words(multiplex, multi_server):
 
 
 @command(internals=True, pass_msg=True)
-def known(multiplex, multi_server, msg=None):
+def known(multiplex: bool, multi_server: str, msg: str = "") -> str:  # pylint:disable=unused-argument
     message = msg.split()[1:]
     logger.info(message)
     ret = requests.get(multi_server + "known?word={}".format(message[0]))
@@ -46,10 +46,10 @@ def known(multiplex, multi_server, msg=None):
 
 
 @command()
-def blap():
+def blap() -> str:
     return "https://sexyferret.science/blap/its_good_to_be_sad.gif"
 
 
 @command()
-def spray():
+def spray() -> str:
     return "https://sexyferret.science/blap/spray.gif"