+++ pyborg/pyborg/mod/mod_discord.py
@@ -1,19 +1,22 @@
-import re
+"""
+pyborg discord module
+"""
+import asyncio
 import logging
+import re
 from functools import partial
+from types import ModuleType
+from typing import Any, Callable, Dict, List, MutableMapping, Optional, Union
 
+import aiohttp
+import attr
 import discord
-import pyborg.commands
-import requests
 import toml
 import venusian
-import attr
 
+from .. import pyborg as pyb_core
+import pyborg.commands as builtin_commands
 from pyborg.util.awoo import normalize_awoos
-from typing import Dict, Union, List, Callable
-from types import ModuleType
-
-# https://github.com/Rapptz/discord.py/blob/master/discord/ext/commands/bot.py#L146
 
 logger = logging.getLogger(__name__)
 
@@ -26,13 +29,14 @@ class PyborgDiscord(discord.Client):
     multi_port: int = attr.ib(default=2001)
     multiplexing: bool = attr.ib(default=True)
     multi_server: str = attr.ib(default="localhost")
+    multi_protocol: str = attr.ib(default="http")
     registry = attr.ib(default=attr.Factory(lambda self: Registry(self), takes_self=True))
     aio_session: aiohttp.ClientSession = attr.ib(init=False)
-    save_status_count : int = attr.ib(default=0, init=False)
-    pyborg = attr.ib(default=None)
-    scanner = attr.ib(default=None)
-    loop = attr.ib(default=None)
-    settings: Dict = attr.ib(default=None)
+    save_status_count: int = attr.ib(default=0, init=False)
+    pyborg: Optional[pyb_core.pyborg] = attr.ib(default=None)
+    scanner: venusian.Scanner = attr.ib(default=None)
+    loop: Optional[asyncio.BaseEventLoop] = attr.ib(default=None)
+    settings: MutableMapping[str, Any] = attr.ib(default=None)
     prefix: str = atrr.ib(default="!")
 
     def __attrs_post_init__(self) -> None:
@@ -50,12 +54,21 @@ class PyborgDiscord(discord.Client):
             raise NotImplementedError
         else:
             self.pyborg = None
-        super().__init__()
+        self.aio_session = aiohttp.ClientSession()
+        super().__init__(loop=self.loop)  # this might create a asyncio.loop!
 
     def our_start(self) -> None:
+        "launch discord.Client main event loop (calls Client.run)"
+
         self.scan()
         if 'token' in self.settings['discord']:
             self.run(self.settings['discord']['token'])
+        else:
+            logger.error("No Token. Set one in your conf file.")
+
+    async def fancy_login(self) -> None:
+        if 'token' in self.settings['discord']:
+            await self.login(self.settings['discord']['token'])
         else:
             logger.error("No Token. Set one in your conf file.")
 
@@ -73,16 +86,17 @@ class PyborgDiscord(discord.Client):
         # s[s.find("<:"):s.find(">")+1] the general idea here
         start = msg.find("<:")
         attempted_emoji = msg[start + 2: msg.find(":", start + 2)]
-        logger.info("_extract_emoji:attempting to emoji: %s", attempted_emoji)
+        logger.debug("_extract_emoji:attempting to emoji: %s", attempted_emoji)
         if attempted_emoji in server_emojis:
             # now replace the range from start to end with the extracted emoji
             end = msg.find(">", start) + 1
             before, _, after = msg.partition(msg[start:end])
             logger.debug(_)
             incoming_message = before + attempted_emoji + after
-            logger.info(incoming_message)
+            logger.debug(incoming_message)
             return incoming_message
         else:
+            # this can be a fancy nitro emoji.
             logger.info("_extract_emoji:someone did a fucky wucky")
             logger.debug("_extract_emoji:OOPSIE WOOPSIE!! Uwu We made a fucky wucky!! A wittle fucko boingo! The code monkeys at our headquarters are working VEWY HAWD to fix this!")
             return msg
@@ -100,34 +114,43 @@ class PyborgDiscord(discord.Client):
             else:
                 if command_name in self.registry.registered:
                     command = self.registry.registered[command_name]
-                    logger.info("Running command %s", command)
-                    logger.info("pass message?: %s", command.pass_msg)
+                    logger.debug("cmd: Running command %s", command)
+                    logger.debug("cmd: pass message?: %s", command.pass_msg)
                     if command.pass_msg:
                         await message.channel.send(command(msg=message.content))
                     else:
                         await message.channel.send(command())
         if message.author == self.user:
-            logger.info("Not learning/responding to self")
+            logger.debug("Not learning/responding to self")
             return
+
+        if self.save_status_count % 5:
+            async with self.aio_session.get(f"http://{self.multi_server}:{self.multi_port}/meta/status.json", raise_for_status=True) as ret_status:
+                data = await ret_status.json()
+                if data["status"]:
+                    await self.change_presence(activity=discord.Game("Saving brain..."))
+                else:
+                    await self.change_presence(activity=discord.Game("hack the planet"))
+
+        self.save_status_count += 1
 
         # Custom Emoji handling here
         # DEBUG:pyborg.mod.mod_discord:<:weedminion:392111795642433556>
         # is this cached? is this fast? who the fuck knows
         if "<:" in message.content:
-            e = message.guild.emojis
-            server_emojis = [x.name for x in e]
+            emojis_possible = message.guild.emojis
+            server_emojis = [x.name for x in emojis_possible]
             logger.debug("got server emojis as: %s", str(server_emojis))
             incoming_message = self._extract_emoji(message.content, server_emojis)
-
         else:
             incoming_message = message.content
 
         # Strip nicknames for pyborg
         line_list = list()
-        for x in incoming_message.split():
-            if x.startswith("<@!"):
-                x = "#nick"
-            line_list.append(x)
+        for chunk in incoming_message.split():
+            if chunk.startswith("<@!"):
+                chunk = "#nick"
+            line_list.append(chunk)
 
         line = " ".join(line_list)
         try:
@@ -141,16 +164,17 @@ class PyborgDiscord(discord.Client):
         line = normalize_awoos(line)
 
         if self.settings['discord']['learning']:
-            self.learn(line)
+            await self.learn(line)
 
         if self.user.mentioned_in(message) or self._plaintext_name(message):
             async with message.channel.typing():
-                msg = self.reply(line)
-                logger.debug("on message: %s" % msg)
+                msg = await self.reply(line)
+                logger.debug("on message: %s", msg)
                 if msg:
                     logger.debug("Sending message...")
                     # if custom emoji: replace to <:weedminion:392111795642433556>
                     # message.server map to full custom emoji
+                    # TODO: measure time spent here?
                     emoji_map = {x.name: x for x in message.guild.emojis}
                     for word in msg.split():
                         if word in emoji_map:
@@ -171,46 +195,46 @@ class PyborgDiscord(discord.Client):
         except KeyError:
             return False
 
-    def learn(self, body: str) -> None:
+    async def learn(self, body: str) -> None:
         """thin wrapper for learn to switch to multiplex mode"""
         if self.settings['pyborg']['multiplex']:
-            ret = requests.post("http://{}:{}/learn".format(self.multi_server, self.multi_port), data={"body": body})
-            if ret.status_code > 499:
-                logger.error("Internal Server Error in pyborg_http. see logs.")
-            else:
-                ret.raise_for_status()
+            await self.aio_session.post(f"http://{self.multi_server}:{self.multi_port}/learn", data={"body": body}, raise_for_status=True)
 
-    def reply(self, body: str) -> Union[str, None]:
-        """thin wrapper for reply to switch to multiplex mode"""
+    async def reply(self, body: str) -> Union[str, None]:
+        """thin wrapper for reply to switch to multiplex mode: now coroutine"""
         if self.settings['pyborg']['multiplex']:
-            ret = requests.post("http://{}:{}/reply".format(self.multi_server, self.multi_port), data={"body": body})
-            if ret.status_code == requests.codes.ok:
-                reply = ret.text
+            url = f"http://{self.multi_server}:{self.multi_port}/reply"
+            async with self.aio_session.post(url, data={"body": body}, raise_for_status=True) as ret:
+                reply = await ret.text()
                 logger.debug("got reply: %s", reply)
-            elif ret.status_code > 499:
-                logger.error("Internal Server Error in pyborg_http. see logs.")
-                return None
-            else:
-                ret.raise_for_status()
             return reply
         else:
             raise NotImplementedError
 
-    def teardown(self) -> None:
-        pass
+    async def teardown(self) -> None:
+        "turn off the bot"
+        await self.aio_session.close()
 
-    def scan(self, module: ModuleType = pyborg.commands) -> None:
+    def scan(self, module: ModuleType = builtin_commands) -> None:
+        "look for commands to add to registry"
         self.scanner = venusian.Scanner(registry=self.registry)
         self.scanner.scan(module)
+#
+#class FancyCallable(Callable):
+#    pass_msg: bool
 
 
-class Registry(object):
+class Registry():
     """Command registry of decorated pyborg commands"""
     def __init__(self, mod: PyborgDiscord) -> None:
         self.registered: Dict[str, Callable] = {}
         self.mod = mod
 
+    def __str__(self):
+        return f"{self.mod} command registry with {len(self.registered.keys())}"
+
     def add(self, name: str, ob: Callable, internals: bool, pass_msg: bool) -> None:
+        "add command to the registry. takes two config options."
         self.registered[name] = ob
         if internals:
             self.registered[name] = partial(ob, self.mod.multiplexing, multi_server="http://{}:{}/".format(self.mod.multi_server, self.mod.multi_port))