+++ pyborg/test/test_reddit.py
@@ -1,9 +1,6 @@
 import unittest
+from unittest import mock
 
-try:
-    from unittest import mock
-except ImportError:
-    import mock
 import pyborg