+++ pyborg/test/test_twitter.py
@@ -1,0 +1,7 @@
+import unittest
+from unittest import mock
+
+
+class TestIgnoreImageTweets(unittest.TestCase):
+    def test_ignore_blank_description(self):
+        pass