+++ pyborg/test/test_upgrade.py
@@ -1,10 +1,7 @@
 import logging
 import unittest
+from unittest import mock
 
-try:
-    from unittest import mock
-except ImportError:
-    import mock
 from click.testing import CliRunner
 
 from pyborg_entrypoint import upgrade_to_json